discord-bot/requirements.txt
discord.py>=2.0
aiohttp>=3.8
flask>=3.0
